Refactor - remove all initialisations of GOBModel from GOB-Core and expect GOBModel as parameter. In preparation for legacy mode
